Move CostOverridesList to own file
